Simulator start on clicking level button

+++ src/main.js
@@ -32,7 +32,8 @@ function createWindow () {
     // Dereference the window object, usually you would store windows
     // in an array if your app supports multi windows, this is the time
     // when you should delete the corresponding element.
-    process.kill(childId)
+    if (childId)
+      process.kill(childId)
     mainWindow = null
   })
 }
@@ -47,7 +48,8 @@ app.on('window-all-closed', function () {
   // On OS X it is common for applications and their menu bar
   // to stay active until the user quits explicitly with Cmd + Q
   if (process.platform !== 'darwin') {
-    process.kill(childId)
+    if (childId)
+      process.kill(childId)
     app.quit()
   }
 })
